rebase timer into component
