+++ src/components/RebaseTimer/rebasetimer.scss
@@ -1,0 +1,12 @@
+.rebase-timer {
+    position: relative;
+    display: inline-block;
+    margin: 10px;
+    min-width: 130px;
+    font-size: 16px;
+    line-height: 22px;
+    p {
+        margin: 0;
+        color: #eee;
+    }
+}
