+++ src/components/Sidebar/sidebar.scss
@@ -8,8 +8,8 @@
     display:flex;
     flex-flow:column;
     background-color:#212121;
-    opacity:0.8;
-    box-shadow:inset 0px 0px 50px black;
+    // opacity:0.8;
+    // box-shadow:inset 0px 0px 50px black;
   }
 
   .dapp-nav {
@@ -37,6 +37,7 @@
     //   font-weight: bold;
     // }
   }
+
 
   .dapp-menu-social {
     width: 100%;
