+++ src/components/TopBar/TopBar.jsx
@@ -3,11 +3,11 @@ import { StaticJsonRpcProvider, Web3Provider } from "@ethersproject/providers";
 import { useSelector, useDispatch } from 'react-redux';
 import Web3Modal from "web3modal";
 import "./topbar.scss";
-import { shorten, trim, getRebaseBlock, secondsUntilBlock, prettifySeconds } from '../../helpers';
+import { shorten } from '../../helpers';
+import RebaseTimer from '../RebaseTimer/RebaseTimer.jsx';
 
 
 function TopBar({ web3Modal, loadWeb3Modal, logoutOfWeb3Modal,  route, setRoute, address, mainnetProvider, blockExplorer }) {
-	const currentBlock  = useSelector((state ) => { return state.app.currentBlock });
 
   const modalButtons = [];
   if (web3Modal) {
@@ -22,20 +22,10 @@ function TopBar({ web3Modal, loadWeb3Modal, logoutOfWeb3Modal,  route, setRoute,
     }
   }
 
-	const timeUntilRebase = () => {
-    if (currentBlock) {
-      const rebaseBlock = getRebaseBlock(currentBlock);
-      const seconds     = secondsUntilBlock(currentBlock, rebaseBlock);
-      return prettifySeconds(seconds);
-    }
-  }
-
   return (
     <div className="dapp-topbar">
-				<div className="rebase-timer">
-					<p>Next Rebase</p>
-					<p>{timeUntilRebase() || "04:20:69"}</p>
-				</div>
+
+				<RebaseTimer />
 
 				<div className="wallet-menu">
 					{address && <a href={`https://etherscan.io/address/${address}`} target="_blank">
