+++ src/components/TopBar/topbar.scss
@@ -8,23 +8,7 @@
     width: 100%;
     height: 70px;
     background-color:#212121;
-    box-shadow:inset 0px 0px 40px black;
-    opacity:0.8;
-}
-
-.rebase-timer {
-    position: relative;
-    display: inline-block;
-    width: 120px;
-    height: 100%;
-    margin-right: 15px;
-    font-size: 16px;
-    padding: 8px;
-    line-height: 22px;
-    p {
-        margin: 0;
-        color: #fff;
-    }
+    padding: 5px;
 }
 
 .wallet-menu {
